Make nimbus-admin tests more thorough.

+++ tests/cc-nimbusAdminList-test.py
@@ -3,10 +3,12 @@
 import pexpect
 import sys
 import os
+import re
 
 to=90
 cc_home=os.environ['CLOUD_CLIENT_HOME']
 nimbus_home=os.environ['NIMBUS_HOME']
+nimbus_user=os.environ['NIMBUS_TEST_USER']
 logfile = sys.stdout
 
 try:
@@ -14,6 +16,10 @@ try:
 except:
 	print "The directory already exists"
 	pass
+
+cmd = "%s/bin/nimbus-list-users %%" % (nimbus_home)
+(x, rc)=pexpect.run(cmd, withexitstatus=1)
+print x
 
 cmd = "%s/bin/cloud-client.sh --transfer --sourcefile /etc/group" % (cc_home)
 (x, rc)=pexpect.run(cmd, withexitstatus=1)
@@ -30,19 +36,19 @@ if rc != 0:
     print "run"
     sys.exit(1)
 
-cmd = "%s/bin/nimbus-admin --list --user user1" % (nimbus_home)
+cmd = "%s/bin/nimbus-admin --list --user %s" % (nimbus_home, nimbus_user)
 print cmd
 (x, rc)=pexpect.run(cmd, withexitstatus=1)
 print x
-if rc != 0:
+if rc != 0 or not re.match(".*id\s*?:\s*?\d.*", x):
     print "error"
     sys.exit(1)
 
-cmd = "%s/bin/nimbus-admin --list --dn /O=Auto/OU=CA/CN=nimbus@" % (nimbus_home)
+cmd = "%s/bin/nimbus-admin --list --dn /O=Auto/OU=CA/CN=%s" % (nimbus_home, nimbus_user)
 print cmd
 (x, rc)=pexpect.run(cmd, withexitstatus=1)
 print x
-if rc != 0:
+if rc != 0 or not re.match(".*id\s*?:\s*?\d.*", x):
 	print "error"
 	sys.exit(1)
 
@@ -50,7 +56,7 @@ cmd = "%s/bin/nimbus-admin --list --host localhost" % (nimbus_home)
 print cmd
 (x, rc)=pexpect.run(cmd, withexitstatus=1)
 print x
-if rc != 0:
+if rc != 0 or not re.match(".*id\s*?:\s*?\d.*", x):
     print "error"
     sys.exit(1)
 
@@ -58,7 +64,7 @@ cmd = "%s/bin/nimbus-admin --list --gid 1" % (nimbus_home)
 print cmd
 (x, rc)=pexpect.run(cmd, withexitstatus=1)
 print x
-if rc != 0:
+if rc != 0 or not re.match(".*id\s*?:\s*?\d.*", x):
     print "error"
     sys.exit(1)
 
@@ -66,7 +72,7 @@ cmd = "%s/bin/nimbus-admin --list --gname UNLIMITED" % (nimbus_home)
 print cmd
 (x, rc)=pexpect.run(cmd, withexitstatus=1)
 print x
-if rc != 0:
+if rc != 0 or not re.match(".*id\s*?:\s*?\d.*", x):
     print "error"
     sys.exit(1)
 
@@ -74,7 +80,7 @@ cmd = "%s/bin/nimbus-admin --list" % (nimbus_home)
 print cmd
 (x, rc)=pexpect.run(cmd, withexitstatus=1)
 print x
-if rc != 0:
+if rc != 0 or not re.match(".*id\s*?:\s*?\d.*", x):
     print "error"
     sys.exit(1)
 
@@ -96,6 +102,12 @@ print cmd
 print x
 if rc != 0:
     print "error"
+    sys.exit(1)
+if not re.search("node\s*:\s*localhost", x):
+    print "not showing localhost node?"
+    sys.exit(1)
+if not re.search("id\s*:\s*\d*,\s\d*", x):
+    print "not showing two vms on localhost node?"
     sys.exit(1)
 
 cmd = "%s/bin/nimbus-admin --batch --shutdown --all" % (nimbus_home)
